+++ src/main/java/com/aksps/BillWise/dto/response/InvoiceItemResponse.java
@@ -1,15 +1,11 @@
-// What it does: DTO representing a single line item's details in the final Invoice response.
-// Why needed: Separated for cleaner code organization.
 package com.aksps.BillWise.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/**
- * DTO representing a single line item's details in the final Invoice response.
- * Separated for cleaner code organization.
- */
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -17,6 +13,6 @@ public class InvoiceItemResponse {
     private String productName;
     private String productSku;
     private Integer quantitySold;
-    private Double unitPriceAtSale;
-    private Double lineTotal;
+    private BigDecimal unitPriceAtSale;
+    private BigDecimal lineTotal;
 }
